feat: add more tests and optimise
