+++ .gitignore
@@ -202,6 +202,7 @@ FakesAssemblies/
 # Zipped Geocoding Data
 geocoding.zip
 testgeocoding.zip
+/resources/carrier.zip
 
 # Performance tests artifacts
 csharp/PhoneNumbers.PerformanceTest/BenchmarkDotNet.Artifacts/
