+++ appveyor.yml
@@ -25,6 +25,7 @@ branches:
 before_build:
   - dotnet restore csharp -s https://api.nuget.org/v3/index.json
   - ps: Compress-Archive -Path "resources\geocoding\*" -DestinationPath "resources\geocoding.zip"
+  - ps: Compress-Archive -Path "resources\carrier\*" -DestinationPath "resources\carrier.zip"
   - ps: Compress-Archive -Path "resources\test\geocoding\*" -DestinationPath "resources\test\testgeocoding.zip"
 build_script:
   - dotnet pack -c Release csharp\PhoneNumbers
